moves examples under the example directory, slack detailed examples to /examples/details, updates references in docs
