+++ docs/readme-ciscospark.md
@@ -35,7 +35,7 @@ ngrok http 3000
 4) Run your bot application using the access token you received, the base url of your bot application, and a secret which is used to validate the origin of incoming webhooks:
 
 ```
-access_token=<MY ACCESS TOKEN> public_address=<https://my_bot_url> secret=<my_secret_phrase> node spark_bot.js
+access_token=<MY ACCESS TOKEN> public_address=<https://my_bot_url> secret=<my_secret_phrase> node examples/spark_bot.js
 ```
 
 5) Your bot should now come online and respond to requests! Find it in Cisco Spark by searching for it's name.
