+++ docs/readme-slack.md
@@ -39,7 +39,7 @@ Copy the API token that Slack gives you. You'll need it.
 4) Run the example bot app, using the token you just copied:
 ​
 ```
-token=REPLACE_THIS_WITH_YOUR_TOKEN node slack_bot.js
+token=REPLACE_THIS_WITH_YOUR_TOKEN node examples/slack_bot.js
 ```
 ​
 5) Your bot should be online! Within Slack, send it a quick direct message to say hello. It should say hello back!
