+++ examples/slack/middleware_example.js
@@ -69,7 +69,7 @@ if (!process.env.token) {
     process.exit(1);
 }
 
-var Botkit = require('../lib/Botkit.js');
+var Botkit = require('../../lib/Botkit.js');
 var os = require('os');
 
 var controller = Botkit.slackbot({
