+++ examples/slack/sentiment_analysis.js
