+++ examples/slack_bot.js
@@ -6,10 +6,11 @@
             \/_____/   \/_____/     \/_/   \/_/\/_/   \/_/     \/_/
 
 
-This is a sample Console bot built with Botkit.
+This is a sample Slack bot built with Botkit.
 
 This bot demonstrates many of the core features of Botkit:
 
+* Connect to Slack using the real time API
 * Receive messages based on "spoken" patterns
 * Reply to messages
 * Use the conversation system to ask questions
@@ -18,11 +19,17 @@ This bot demonstrates many of the core features of Botkit:
 
 # RUN THE BOT:
 
+  Get a Bot token from Slack:
+
+    -> http://my.slack.com/services/new/bot
+
   Run your bot from the command line:
 
-    node console_bot.js
+    token=<MY TOKEN> node slack_bot.js
 
 # USE THE BOT:
+
+  Find your bot inside Slack to send it a direct message.
 
   Say: "Hello"
 
@@ -56,16 +63,35 @@ This bot demonstrates many of the core features of Botkit:
 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
 
-var Botkit = require('./lib/Botkit.js');
+
+if (!process.env.token) {
+    console.log('Error: Specify token in environment');
+    process.exit(1);
+}
+
+var Botkit = require('../lib/Botkit.js');
 var os = require('os');
 
-var controller = Botkit.consolebot({
-    debug: false,
+var controller = Botkit.slackbot({
+    debug: true,
 });
 
-var bot = controller.spawn();
+var bot = controller.spawn({
+    token: process.env.token
+}).startRTM();
 
-controller.hears(['hello', 'hi'], 'message_received', function(bot, message) {
+controller.hears(['hello', 'hi'], 'direct_message,direct_mention,mention', function(bot, message) {
+
+    bot.api.reactions.add({
+        timestamp: message.ts,
+        channel: message.channel,
+        name: 'robot_face',
+    }, function(err, res) {
+        if (err) {
+            bot.botkit.log('Failed to add emoji reaction :(', err);
+        }
+    });
+
 
     controller.storage.users.get(message.user, function(err, user) {
         if (user && user.name) {
@@ -76,7 +102,7 @@ controller.hears(['hello', 'hi'], 'message_received', function(bot, message) {
     });
 });
 
-controller.hears(['call me (.*)', 'my name is (.*)'], 'message_received', function(bot, message) {
+controller.hears(['call me (.*)', 'my name is (.*)'], 'direct_message,direct_mention,mention', function(bot, message) {
     var name = message.match[1];
     controller.storage.users.get(message.user, function(err, user) {
         if (!user) {
@@ -91,7 +117,7 @@ controller.hears(['call me (.*)', 'my name is (.*)'], 'message_received', functi
     });
 });
 
-controller.hears(['what is my name', 'who am i'], 'message_received', function(bot, message) {
+controller.hears(['what is my name', 'who am i'], 'direct_message,direct_mention,mention', function(bot, message) {
 
     controller.storage.users.get(message.user, function(err, user) {
         if (user && user.name) {
@@ -160,7 +186,7 @@ controller.hears(['what is my name', 'who am i'], 'message_received', function(b
 });
 
 
-controller.hears(['shutdown'], 'message_received', function(bot, message) {
+controller.hears(['shutdown'], 'direct_message,direct_mention,mention', function(bot, message) {
 
     bot.startConversation(message, function(err, convo) {
