+++ examples/twilio_ipm_bot.js
@@ -1,4 +1,4 @@
-var Botkit = require('./lib/Botkit.js');
+var Botkit = require('../lib/Botkit.js');
 var os = require('os');
 var controller = Botkit.twilioipmbot({
     debug: false,
